Refactor: move NonCopyable to core
