+++ src/core/include/core/NonCopyable.hpp
